fix: support float score in DB

+++ server/db/migration/000003_add_ratings_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE ratings (
     id UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
     user_id UUID REFERENCES users(id),
     movie_id UUID REFERENCES movies(id),
-    score INTEGER NOT NULL,
+    score REAL NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP
 );
